Move interpolation tests to own module
